requirements.txt
about-time==4.2.1
absl-py==1.4.0
adafruit-ampy==1.1.0
aiohttp==3.8.3
aiosignal==1.3.1
alabaster==0.7.13
alive-progress==3.0.1
anyio==3.6.2
anytree==2.8.0
appdirs==1.4.4
apturl==0.5.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
astroid==2.15.5
asttokens==2.2.1
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
azure-core==1.27.1
azure-storage-blob==12.16.0
Babel==2.11.0
backcall==0.2.0
bcrypt==3.1.7
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.6.2
breathe==4.35.0
Brlapi==0.7.0
bs4==0.0.1
cachetools==5.3.1
canopen==2.1.0
capstone==4.0.2
catppuccin-matplotlib==0.4
cbor==1.0.0
cbor2==5.4.6
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
chrome-gnome-shell==0.0.0
clang-format==16.0.6
click==8.1.3
cmsis-pack-manager==0.5.2
colorama==0.4.3
comm==0.1.2
command-not-found==0.3
commonmark==0.9.1
contextlib2==21.6.0
contourpy==1.0.7
coverage==7.2.7
cryptography==2.8
cssselect==1.2.0
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.16
debugpy==1.6.5
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.6
discord==2.1.0
discord.py==2.1.0
distlib==0.3.6
distro==1.4.0
distro-info===0.23ubuntu1
dlib==19.24.0
dnspython==2.3.0
docopt==0.6.2
docutils==0.20.1
duplicity==0.8.12.0
ecdsa==0.18.0
email-validator==2.0.0.post2
entrypoints==0.3
exceptiongroup==1.1.0
executing==1.2.0
face-recognition==1.3.0
face-recognition-models==0.3.0
fake-useragent==1.1.1
fastapi==0.95.1
fasteners==0.14.1
fastjsonschema==2.16.2
filelock==3.9.0
filterpy==1.4.5
Flask==2.3.2
flatbuffers==23.5.26
fonttools==4.33.3
frozenlist==1.3.3
future==0.18.2
gast==0.4.0
gcovr==6.0
getmac==0.9.4
gitlint==0.19.1
gitlint-core==0.19.1
google-auth==2.19.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grapheme==0.6.0
graphviz==0.20.1
greenlet==0.4.15
grip==4.6.1
grpcio==1.54.2
grpcio-tools==1.54.2
h11==0.14.0
h5py==3.8.0
html5lib==1.0.1
httpcore==0.17.2
httplib2==0.14.0
httptools==0.5.0
httpx==0.24.1
idna==2.8
imagesize==1.4.1
imgtool==1.10.0
importlib-metadata==6.0.0
importlib-resources==5.10.2
iniconfig==2.0.0
intelhex==2.3.0
intervaltree==3.1.0
ipykernel==6.20.2
ipython==8.8.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isodate==0.6.1
isort==5.12.0
itsdangerous==2.1.2
jax==0.4.12
jedi==0.18.2
Jinja2==3.1.2
joblib==1.1.0
json5==0.9.11
jsonschema==4.17.3
junit2html==30.1.3
junitparser==3.1.0
jupyter==1.0.0
jupyter-client==7.4.1
jupyter-console==6.4.4
jupyter-core==5.1.3
jupyter-events==0.6.3
jupyter-server==1.23.5
jupyter-server-terminals==0.4.4
jupyterlab==3.5.2
jupyterlab-pygments==0.2.2
jupyterlab-server==2.19.0
jupyterlab-widgets==3.0.5
keras==2.12.0
keyring==18.0.1
kiwisolver==1.4.2
language-selector==0.1
lark==1.1.5
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.9.0
libclang==16.0.0
libusb-package==1.0.26.2
lockfile==0.12.2
louis==3.12.0
lpc-checksum==2.2.0
lxml==4.9.2
m2r2==0.3.3.post2
macaroonbakery==1.3.1
Mako==1.1.0
mariadb==1.1.5.post3
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.5.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mistune==0.8.4
ml-dtypes==0.2.0
mock==5.0.2
monotonic==1.5
msgpack==1.0.5
multidict==6.0.4
mypy==1.3.0
mypy-extensions==1.0.0
mysql==0.0.3
mysql-connector-python==8.0.32
mysqlclient==2.1.1
natsort==8.3.1
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.8
nbformat==5.7.3
nest-asyncio==1.5.6
netifaces==0.10.4
networkx==2.8.3
nltk==3.7
nose==1.3.7
notebook==6.5.2
notebook-shim==0.2.2
numpy==1.22.4
oauthlib==3.1.0
olefile==0.46
opencv-python==4.7.0.72
opt-einsum==3.3.0
orjson==3.9.1
outcome==1.2.0
packaging==21.3
paho-mqtt==1.6.1
pandas==1.5.3
pandocfilters==1.5.0
paramiko==2.6.0
parse==1.19.0
parso==0.8.3
path-and-address==2.0.1
pathspec==0.11.1
pbr==5.11.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
pkgutil-resolve-name==1.3.10
platformdirs==2.6.2
pluggy==1.0.0
ply==3.11
prettytable==3.8.0
progress==1.6
progress-bar==8
prometheus-client==0.15.0
prompt-toolkit==3.0.36
protobuf==4.23.3
psutil==5.5.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyaml==21.10.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycairo==1.16.2
pycparser==2.21
pycrypto==2.6.1
pycups==1.9.73
pydantic==1.10.7
pyee==8.2.2
pyelftools==0.29
pygame==2.1.2
PyGithub==1.58.2
Pygments==2.14.0
PyGObject==3.36.0
PyJWT==2.7.0
pykwalify==1.8.0
pylink-square==1.1.0
pylint==2.17.4
pymacaroons==0.13.0
PyNaCl==1.5.0
pynvim==0.4.1
pyocd==0.35.1
pyparsing==3.0.9
pyppeteer==1.0.2
pyquery==2.0.0
pyRFC3339==1.1
pyrsistent==0.19.3
pyserial==3.5
PySocks==1.7.1
pytest==7.3.2
python-apt==2.0.1+ubuntu0.20.4.1
python-barcode==0.14.0
python-can==4.2.2
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-dotenv==0.20.0
python-json-logger==2.0.4
python-magic==0.4.27
python-multipart==0.0.6
pytz==2022.7.1
pyusb==1.2.1
pyxattr==0.6.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==23.0.0
qtconsole==5.4.0
QtPy==2.3.0
recommonmark==0.6.0
regex==2022.6.2
reportlab==3.5.34
requests==2.28.2
requests-html==0.10.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
scapy==2.5.0
schema==0.7.5
scipy==1.10.0
scour==0.37
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
selenium==4.7.2
Send2Trash==1.8.0
sh==2.0.4
simplejson==3.16.0
six==1.14.0
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sphinx==5.3.0
sphinx-copybutton==0.5.2
sphinx-markdown-tables==0.0.17
sphinx-ncs-theme==1.0.17
sphinx-notfound-page==0.8.3
sphinx-rtd-theme==1.2.2
sphinx-tabs==3.4.1
sphinx-togglebutton==0.3.2
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-mscgen==0.6
sphinxcontrib-plantuml==0.25
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sphinxcontrib-svg2pdfconverter==1.2.2
spline==1.12
sql==2022.4.0
ssh-import-id==5.10
stack-data==0.6.2
starlette==0.26.1
systemd-python==234
tabulate==0.9.0
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
terminado==0.17.1
testresources==2.0.1
tinycss2==1.2.1
tomli==2.0.1
tomlkit==0.11.8
tornado==6.2
tqdm==4.64.0
traitlets==5.8.1
trio==0.22.0
trio-websocket==0.9.2
typing-extensions==4.5.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
ujson==5.8.0
unattended-upgrades==0.1
urllib3==1.26.14
urwid==2.0.1
usb-creator==0.3.7
uvicorn==0.22.0
uvloop==0.17.0
virtualenv==20.19.0
voila==0.4.0
w3lib==2.1.1
wadllib==1.3.3
watchfiles==0.19.0
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.4.2
websockets==10.4
Werkzeug==2.3.4
west==1.1.0
who-is-on-my-wifi==1.3.5
widgetsnbextension==4.0.5
wrapt==1.14.1
wsproto==1.2.0
wxPython==4.2.0
xkit==0.0.0
yamllint==1.32.0
yarl==1.8.2
youtube-dl==2020.3.24
zcbor==0.7.0
zipp==3.11.0
